+++ protected/humhub/components/captcha/YiiCaptchaInput.php
@@ -17,10 +17,36 @@ use yii\captcha\Captcha;
 class YiiCaptchaInput extends Captcha
 {
     public $captchaAction = '/captcha/yii';
+    /**
+     * The form input HTML element (e.g. #my-text-field) that needs to be focused to show the Captcha input
+     * If empty, the Captcha input is always displayed
+     * @since 1.18.1
+     */
+    public ?string $showOnFocusElement = null;
 
     public function init()
     {
         $this->options['placeholder'] = Yii::t('base', 'Enter security code above');
         parent::init();
+
+        $id = $this->options['id'] ?? null;
+        if (!$id) {
+            $this->showOnFocusElement = null;
+        }
+
+        if ($this->showOnFocusElement) {
+            $view = $this->getView();
+            $view->registerJs("
+                $(function () {
+                    const container = $('#$id').parent();
+                    if (!$('#$id.is-invalid').length) {
+                        container.hide();
+                        $('$this->showOnFocusElement').on('focus', function () {
+                            container.fadeIn(500);
+                        });
+                    }
+                });
+            ");
+        }
     }
 }