+++ protected/humhub/modules/user/views/auth/login.php
@@ -102,9 +102,9 @@ $this->pageTitle = Yii::t('UserModule.auth', 'Login');
                 <?php $form = ActiveForm::begin(['id' => 'invite-form']); ?>
                 <?= $form->field($invite, 'email')->input('email', ['id' => 'register-email', 'placeholder' => $invite->getAttributeLabel('email'), 'aria-label' => $invite->getAttributeLabel('email')])->label(false); ?>
                 <?php if ($invite->showCaptureInRegisterForm()) : ?>
-                    <div id="registration-form-captcha" style="display: none;">
-                        <?= $form->field($invite, 'captcha')->widget(CaptchaField::class)->label(false) ?>
-                    </div>
+                    <?= $form->field($invite, 'captcha')
+                        ->widget(CaptchaField::class, ['showOnFocusElement' => '#register-email'])
+                        ->label(false) ?>
                 <?php endif; ?>
 
                 <?= Html::submitButton(Yii::t('UserModule.auth', 'Register'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
@@ -139,14 +139,4 @@ $this->pageTitle = Yii::t('UserModule.auth', 'Login');
     $('#login-form').removeClass('bounceIn');
     $('#app-title').removeClass('fadeIn');
     <?php } ?>
-
-    <?php if ($invite->showCaptureInRegisterForm()) : ?>
-    if ($('#invite-captcha.is-invalid').length) {
-        $('#registration-form-captcha').show();
-    } else {
-        $('#register-email').on('focus', function () {
-            $('#registration-form-captcha').fadeIn(500);
-        });
-    }
-    <?php endif; ?>
 </script>