+++ protected/humhub/modules/user/views/password-recovery/reset_success.php
@@ -5,7 +5,7 @@ use yii\helpers\Url;
 
 $this->pageTitle = Yii::t('UserModule.auth', 'Password reset');
 ?>
-<div id="user-password-recovery-reset-success" class="container">
+<div id="user-password-recovery-reset-success" class="container container-password">
     <?= SiteLogo::widget(['place' => SiteLogo::PLACE_LOGIN]) ?>
     <br>