+++ CHANGELOG.md
@@ -4,6 +4,7 @@ HumHub Changelog
 1.18.2 (Unreleased)
 ----------------------
 - Fix #8046: Profile Header Title HTML encoded twice since 1.18.1
+- Fix #8043: User display name bottom truncated in top bar
 
 1.18.1 (March 2, 2026)
 ----------------------