+++ CHANGELOG.md
@@ -5,7 +5,7 @@ HumHub Changelog
 ----------------------
 - Fix #8046: Profile Header Title HTML encoded twice since 1.18.1
 - Fix #8043: User display name bottom truncated in top bar
-- Fix #1102: Registration - Display Captcha field if invalid after form submit via keyboard
+- Fix #8051: Registration - Display Captcha field if invalid after form submit via keyboard
 
 1.18.1 (March 2, 2026)
 ----------------------