+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HumHub Changelog
 - Fix #8046: Profile Header Title HTML encoded twice since 1.18.1
 - Fix #8043: User display name bottom truncated in top bar
 - Fix #8051: Registration - Display Captcha field if invalid after form submit via keyboard
+- Enh #8051: Add `AltchaCaptchaInput::$showOnFocusElement` and `YiiCaptchaInput::$showOnFocusElement` (see [migration guide](https://github.com/humhub/humhub/blob/master/MIGRATE-DEV.md#version-1181) for details)
 
 1.18.1 (March 2, 2026)
 ----------------------