+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HumHub Changelog
 ----------------------
 - Fix #8046: Profile Header Title HTML encoded twice since 1.18.1
 - Fix #8043: User display name bottom truncated in top bar
+- Fix #8054: Login layout widths (Default: Bootstrap width, Registration: 500px, Login & Password: 300px, Login with multiple SSO buttons: 500px)
 
 1.18.1 (March 2, 2026)
 ----------------------